phyui: refactor python
